Set backspace behaviour
On some systems, you may experience problems using the backspace or
delete keys. This tip discusses the causes and solutions. Generally,
these problems arise on Unix-based systems because of the wide variety
of hardware and software involved.

Backspace works in Insert mode (e.g. not inserting a ^?), but won't
delete over line breaks, or automatically-inserted indentation, or the
place where insert mode started. This change fixes this behavior.