Wrap long lines
This option changes how text is displayed. It doesn't change the text in
the buffer. Lines longer than the width of the window will wrap and
displaying continues on the next line. When the cursor is moved to a part
that is not shown, the screen will scroll horizontally.
The line will be broken in the middle of a word if necessary.